refactor: refactoring Header.tsx & me response
