+++ src/features/auth/model/hooks/useRegisterMutation.ts
@@ -24,7 +24,7 @@ export const useRegisterMutation = () => {
     onSuccess: (data) => {
       authActions.setStatus('authenticated');
 
-      queryClient.setQueryData(['auth', 'me'], { user: data.user });
+      queryClient.setQueryData(['auth', 'me'], { user: data.result.user });
 
       router.push(AppRouter.main);
     },
